fix serializable (#11)
suppose there are three transactions trying to execute in parallel.
the first one sets _waitOnCurrentFunction and starts running.
the other two await _waitOnCurrentFunction. when the first transaction finishes, the other two start running, without waiting for each other.

using a while loop is the standard pattern when waiting for a condition variable, so it seems fine to use it here.

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.